src/cfehome/api.py
from ninja import NinjaAPI, Schema
from ninja_extra import NinjaExtraAPI
from ninja_jwt.authentication import JWTAuth
from ninja_jwt.controller import NinjaJWTDefaultController

api = NinjaExtraAPI()
api.register_controllers(NinjaJWTDefaultController)
api.add_router("/waitlist/", "waitlist.api.router")


# api = NinjaAPI()


# User schema imported from Schema from ninja
class UserSchema(Schema):
    username: str
    is_authenticated: bool
    # is not request.user.is_authenticated
    email: str = None


@api.get("/hello")
def hello(request):
    print(request)
    return {"message": "Hello World"}


@api.get("/me", response=UserSchema, auth=JWTAuth())
def me(request):
    return request.user
