requirements.txt
Django
Django-ninja
django-cors-headers
django-ninja-jwt[crypto]
pydantic[email]
rav